Replace pre-build event with linked files
Due to the nature that the packaging works, the files that are copied during the pre-build event are not included on the first build. Any following build will include them. Since on the build server we must assume the source is cleaned before it is cloned, the first build would not include the files. Linking the files doesn't work for the files in the UploadedResources\Extras because the original files are at the same level. This is fixed by creating copies since it's just about 6 files.