Merged PR 4676: M4-70755_Add robot test to verify set logo
This code deals with testing persistence of values and uploading customer logo in Welcome page of ITMC. It also verifies the same in Search page of WC. Three of the tests which are failing are related to Refresh cycle for which https://everyangle.visualstudio.com/EveryAngle/_git/WebServer/pullrequest/4305 PR has already gone into master. Please give the waiver for the same tests.
![image.png](https://everyangle.visualstudio.com/944c3508-25d4-46f4-ad5f-72eb92421f4a/_apis/git/repositories/5d6d255e-4f64-402e-83ea-bdae40b79152/pullRequests/4676/attachments/image%20(2).png)