Merged PR 6685: M4-89716: Fixed conflicting sessions when using multiple EA editions on one server.
M4-89716: Fixed conflicting sessions when using multiple EA editions on one server. This is fixed by storing the authentication cookie on an application-specific path, instead of the root path. Also changed the name of the cookie to 'STSTOKEN' to avoid conflicts with older cookies.

![image (3).png](https://everyangle.visualstudio.com/944c3508-25d4-46f4-ad5f-72eb92421f4a/_apis/git/repositories/5d6d255e-4f64-402e-83ea-bdae40b79152/pullRequests/6685/attachments/image%20%283%29.png)
AppServer failure is not related (also fails on https://nl-cqamb04.everyangle.org/TestServer/)
![image (2).png](https://everyangle.visualstudio.com/944c3508-25d4-46f4-ad5f-72eb92421f4a/_apis/git/repositories/5d6d255e-4f64-402e-83ea-bdae40b79152/pullRequests/6685/attachments/image%20%282%29.png)