Merged PR 6968: M4-92053_Fixing by running the test as single
M4-92053_Fixing by running the test as single since it is failing due to parallel run of **Edit the model verify and restore content back to the same values** test which uses the same session as this test due to SSO. One test failure is not due to current changes and a ticket is already present for the same.
![image.png](https://everyangle.visualstudio.com/944c3508-25d4-46f4-ad5f-72eb92421f4a/_apis/git/repositories/5d6d255e-4f64-402e-83ea-bdae40b79152/pullRequests/6968/attachments/image.png)
![image (2).png](https://everyangle.visualstudio.com/944c3508-25d4-46f4-ad5f-72eb92421f4a/_apis/git/repositories/5d6d255e-4f64-402e-83ea-bdae40b79152/pullRequests/6968/attachments/image%20%282%29.png)
Two test failures are not due to current changes and a ticket is already present for one of them.
![image (3).png](blob:https://everyangle.visualstudio.com/6a819e81-e41b-49da-9a24-94aff01be0a2)
![image (4).png](https://everyangle.visualstudio.com/944c3508-25d4-46f4-ad5f-72eb92421f4a/_apis/git/repositories/5d6d255e-4f64-402e-83ea-bdae40b79152/pullRequests/6968/attachments/image%20%284%29.png)
![image (5).png](https://everyangle.visualstudio.com/944c3508-25d4-46f4-ad5f-72eb92421f4a/_apis/git/repositories/5d6d255e-4f64-402e-83ea-bdae40b79152/pullRequests/6968/attachments/image%20%285%29.png)